+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/NRD_SkillMath.lua
@@ -296,6 +296,7 @@ function ComputeBaseWeaponDamage(weapon)
     local baseDmgFromBase = baseStat.DamageFromBase * 0.01
     local baseMinDamage = baseStat.MinDamage
     local baseMaxDamage = baseStat.MaxDamage
+    Ext.Print("Base:",baseDmgFromBase, "Min:", baseMinDamage, "Max:", baseMaxDamage)
     local damageBoost = 0
 
     for i, stat in pairs(stats) do